+++ composeApp/src/commonMain/kotlin/domain/usecase/ManageOrderUseCase.kt
@@ -9,21 +9,24 @@ class ManageOrderUseCase(
 ) {
     suspend fun getAllPresets(
         outletID: Int, restID: Int,
-        checkId: Int
+        checkId: Long
     ): List<Preset> {
         return orderGateway.getAllPresets(outletID, restID, checkId)
     }
 
     suspend fun getAllItems(
-        outletID: Int, restID: Int, presetID: Int,
-        checkId: Int,
+        outletID: Int,
+        restID: Int,
+        presetID: Int,
+        checkId: Long,
         priceLvlId: Int
     ): List<Item> {
         return orderGateway.getAllItems(outletID, restID, presetID, checkId, priceLvlId)
     }
 
     suspend fun checkItemHasModifiers(
-        restID: Int, itemID: Int,
+        restID: Int,
+        itemID: Int,
         priceLvlId: Int
     ): List<Item> {
         return orderGateway.checkItemHasModifiers(restID, itemID, priceLvlId)
